+++ src/VaultSharp/V1/SecretsEngines/AliCloud/AliCloudSecretsEngineProvider.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using VaultSharp.Core;
 using VaultSharp.V1.Commons;
+using VaultSharp.V1.SecretsEngines.AliCloud.Models;
 
 namespace VaultSharp.V1.SecretsEngines.AliCloud
 {
@@ -12,6 +13,42 @@ namespace VaultSharp.V1.SecretsEngines.AliCloud
         public AliCloudSecretsEngineProvider(Polymath polymath)
         {
             _polymath = polymath;
+        }
+
+        public async Task ConfigureRootCredentialsAsync(CreateRootCredentialsConfigModel createRootCredentialsConfigModel = null, string mountPoint = null)
+        {
+            await _polymath.MakeVaultApiRequest(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/config", HttpMethod.Post, createRootCredentialsConfigModel).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
+        public async Task<Secret<RootCredentialsConfigModel>> ReadRootCredentialsConfigAsync(string mountPoint = null, string wrapTimeToLive = null)
+        {
+            return await _polymath.MakeVaultApiRequest<Secret<RootCredentialsConfigModel>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/config", HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
+        public async Task WriteRoleAsync(string roleName, CreateAliCloudRoleModel createAliCloudRoleModel, string mountPoint = null)
+        {
+            Checker.NotNull(roleName, "roleName");
+
+            await _polymath.MakeVaultApiRequest(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/role/" + roleName.Trim('/'), HttpMethod.Post, requestData: createAliCloudRoleModel).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
+        public async Task<Secret<AliCloudRoleModel>> ReadRoleAsync(string roleName, string mountPoint = null, string wrapTimeToLive = null)
+        {
+            Checker.NotNull(roleName, "roleName");
+
+            return await _polymath.MakeVaultApiRequest<Secret<AliCloudRoleModel>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/role/" + roleName.Trim('/'), HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
+        public async Task<Secret<ListInfo>> ReadAllRolesAsync(string mountPoint = null, string wrapTimeToLive = null)
+        {
+            return await _polymath.MakeVaultApiRequest<Secret<ListInfo>>(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/role", _polymath.ListHttpMethod, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
+        public async Task DeleteRoleAsync(string roleName, string mountPoint = null)
+        {
+            Checker.NotNull(roleName, "roleName");
+
+            await _polymath.MakeVaultApiRequest(mountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.AliCloud, "/role/" + roleName.Trim('/'), HttpMethod.Delete).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
         }
 
         public async Task<Secret<AliCloudCredentials>> GetCredentialsAsync(string aliCloudRoleName, string aliCloudMountPoint = null, string wrapTimeToLive = null)