+++ src/VaultSharp/V1/SecretsEngines/AliCloud/Models/AliCloudInlinePolicyModel.cs
@@ -1,0 +1,14 @@
+﻿
+using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.AliCloud.Models
+{
+    public class AliCloudInlinePolicyModel
+    {
+        [JsonProperty("hash")]
+        public string Hash { get; set; }
+
+        [JsonProperty("policy_document")]
+        public AliCloudInlinePolicyDocumentModel PolicyDocument { get; set; }
+    }
+}