+++ src/VaultSharp/V1/SecretsEngines/AliCloud/Models/AliCloudRemotePolicyModel.cs
@@ -1,0 +1,14 @@
+﻿
+using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.AliCloud.Models
+{
+    public class AliCloudRemotePolicyModel
+    {
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("type")]
+        public string Type { get; set; }
+    }
+}