+++ src/VaultSharp/V1/SecretsEngines/AliCloud/Models/RootCredentialsConfigModel.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.AliCloud.Models
+{
+    public class RootCredentialsConfigModel
+    {
+        [JsonProperty("access_key")]
+        public string AccessKey { get; set; }
+    }
+}